requirements.txt
openai
semantic_kernel
instructor
pydantic
pydantic-settings
redis
fastapi[standard]
streamlit
azure-ai-inference
kubernetes
aiohttp